+++ src/transactions/typings/Transaction.d.ts
@@ -2,9 +2,9 @@ import { Balance, Claims } from '../../wallet';
 import {
   TransactionInput,
   TransactionAttribute,
-  TransactionOutputLike,
+  TransactionOutput,
   Witness,
-} from './core';
+} from './components';
 
 /**
  * Transactions are what you use to interact with the blockchain.
@@ -17,7 +17,7 @@ export class Transaction {
   public version: number
   public attributes: TransactionAttribute[]
   public inputs: TransactionInput[]
-  public outputs: TransactionOutputLike[]
+  public outputs: TransactionOutput[]
   public scripts: Witness[]
 
   constructor(tx: Transaction)
@@ -32,10 +32,10 @@ export class Transaction {
   static createClaimTx(publicKeyOrAddress: string, claimData: Claims, override: object): Transaction
 
   /** Creates a ContractTransaction with the given parameters. */
-  static createContractTx(balances: Balance, intents: TransactionOutputLike[], override: object): Transaction
+  static createContractTx(balances: Balance, intents: TransactionOutput[], override: object): Transaction
 
   /** Creates an InvocationTransaction with the given parameters. */
-  static createInvocationTx(balance: Balance, intents: TransactionOutputLike[], invoke: object | string, gasCost: number, override: object): Transaction
+  static createInvocationTx(balance: Balance, intents: TransactionOutput[], invoke: object | string, gasCost: number, override: object): Transaction
 
   static deserialize(hexstring: string): Transaction
